zoom to fit all orgs working. Created AllOrganizationsMap component to clean the code
